+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ### Fixed
 - check-instance-events.rb: Ignore completed instance-stop ec2 events
 - check-instance-events.rb: Ignore canceled system-maintence ec2 events
+## Added
+- Added check-instance-reachability.rb: looks up all instances from a filter set and pings
 
 ## [3.0.0] - 2016-05-05
 ### Removed